+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+0.8.0 (2016-03-23)
+-----------------
+* Better key mapping and fixes on velocity
+
 0.7.0 (2016-02-03)
 -----------------
 * Added linear speed on Y and Z axis for left/right and up/down movements